refactor: file direction
