+++ src/components/DisplayProductInfo/DisplayProductInfo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/global';
+@import '../../styles/global';
 .wrapper {
   display: flex;
 }
